+++ src/app/components/Setting/CoffeeScriptVersion/index.js
@@ -1,21 +1,14 @@
 import React from 'react'
-import { connect } from 'react-redux'
-import { setCoffeeScriptVersion } from '../../../actions/settings'
+import Selector from '../Selector'
 import { getAllVersions as getAllCoffeeScriptVersions } from '../../../coffeescript'
 
-const CoffeeScriptVersionSetting = ({ dispatch, version }) => {
+export default function ({ version, onChange }) {
   return (
-    <div className='Settings__setting'>
-      <label className='Setting__label'>CoffeeScript Version</label>
-      <select className='Setting__selector' value={version} onChange={e => {
-        dispatch(setCoffeeScriptVersion(e.target.value))
-      }}>
-        {getAllCoffeeScriptVersions().map(v => {
-          return <option value={v} key={v}>{v}</option>
-        })}
-      </select>
-    </div>
+    <Selector
+      label='CoffeeScript Version'
+      initialValue={version}
+      onChange={onChange}
+      options={getAllCoffeeScriptVersions()}
+    />
   )
 }
-
-export default connect()(CoffeeScriptVersionSetting)