+++ src/app/components/Setting/EditorTheme/index.js
@@ -1,21 +1,14 @@
 import React from 'react'
-import { connect } from 'react-redux'
-import { setEditorTheme } from '../../../actions/settings'
+import Selector from '../Selector'
 import { getAllThemes as getAllEditorThemes } from '../../../editor'
 
-const EditorThemeSetting = ({ dispatch, theme }) => {
+export default function ({ theme, onChange }) {
   return (
-    <div className='Settings__setting'>
-      <label className='Setting__label'>Editor Theme</label>
-      <select className='Setting__selector' value={theme} onChange={e => {
-        dispatch(setEditorTheme(e.target.value))
-      }}>
-        {getAllEditorThemes().map(v => {
-          return <option value={v} key={v}>{v}</option>
-        })}
-      </select>
-    </div>
+    <Selector
+      label='Editor Theme'
+      initialValue={theme}
+      onChange={onChange}
+      options={getAllEditorThemes()}
+    />
   )
 }
-
-export default connect()(EditorThemeSetting)