+++ src/app/index.js
@@ -1,9 +1,8 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { Provider } from 'react-redux'
-
 import store from './store'
-import App from './components/App'
+import App from './containers/App'
 
 import 'normalize.css'
 import 'codemirror/lib/codemirror.css'