+++ src/electron/preload.js
@@ -7,13 +7,16 @@ const Menu = remote.Menu
 const InputMenu = Menu.buildFromTemplate([
   {
     type: 'separator'
-  }, {
+  },
+  {
     label: 'Cut',
     role: 'cut'
-  }, {
+  },
+  {
     label: 'Copy',
     role: 'copy'
-  }, {
+  },
+  {
     label: 'Paste',
     role: 'paste'
   }
@@ -21,7 +24,7 @@ const InputMenu = Menu.buildFromTemplate([
 
 // https://github.com/electron/electron/issues/4068#issuecomment-170911307
 window.addEventListener('DOMContentLoaded', () => {
-  document.body.addEventListener('contextmenu', (e) => {
+  document.body.addEventListener('contextmenu', e => {
     e.preventDefault()
     e.stopPropagation()
 
@@ -30,6 +33,7 @@ window.addEventListener('DOMContentLoaded', () => {
     while (node) {
       if (node.nodeName.match(/DIV/) && node.className && node.className.match(/^CodeMirror.*$/i)) {
         InputMenu.popup(remote.getCurrentWindow())
+
         break
       }